# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger (свэг) - OpenAPI 3.0
  description: |-
  version: '1.0'
...
paths:
...
  /loan/tariffs:
    get:
      tags:
        - Loan
      security:
        - bearerAuth: []
      parameters:
        - name: nameQuery
          in: query
          required: false
          schema:
            type: string
        - name: sortingProperty
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/LoanTariffSortingProperty'
        - name: sortingOrder
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SortingOrder'
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список тарифов получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanTariff'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized; возвращается, когда либо вызывающий не авторизован
        '403':
          description: Forbidden; возвращается, когда вызывающий - заблокированный сотрудник
        '500':
          description: Internal server error
  /employee/loan/tariffs/create:
    post:
      tags:
        - Loan
      description: Запрос доступен только сотрудникам (если сотрудник не заблокирован)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanTariffCreationRequest'
      responses:
        '200':
          description: Тариф успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanTariff'
        '401':
          description: Unauthorized; возвращается, когда либо вызывающий не авторизован
        '403':
          description: Forbidden; возвращается, когда либо вызывающий - заблокированный сотрудник или клиент
        '500':
          description: Internal server error
  /employee/loan/tariffs/{tariffId}/delete:
    delete:
      tags:
        - Loan
      description: Запрос доступен только сотрудникам (если сотрудник не заблокирован)
      security:
        - bearerAuth: []
      parameters:
        - name: tariffId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Тариф успешно удалён
        '400':
          description: Bad Request; возвращается, когда тариф с указанным id не существует
        '401':
          description: Unauthorized; возвращается, когда либо вызывающий не авторизован
        '403':
          description: Forbidden; возвращается, когда либо вызывающий - заблокированный сотрудник или клиент
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoanTariff:
      type: object
      required:
        - id
        - name
        - interestRate
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Скам"
        interestRate:
          type: string
          example: "25.5"
        createdAt:
          type: string
          format: date-time
    LoanTariffSortingProperty:
      type: string
      enum:
        - NAME
        - CREATED_AT
        - INTEREST_RATE
    SortingOrder:
      type: string
      enum:
        - ASCENDING
        - DESCENDING
    LoanTariffCreationRequest:
      type: object
      required:
        - name
        - interestRate
      properties:
        name:
          type: string
          example: "Скам"
        interestRate:
          type: string
          example: "25.5"
...